+++ rust/operator-binary/src/crd.rs
@@ -89,6 +89,10 @@ pub struct AutoTlsBackend {
     /// Configures the certificate authority used to issue Pod certificates.
     pub ca: AutoTlsCa,
 
+    /// Additional trust roots which are added to the provided `ca.crt` file.
+    #[serde(default)]
+    pub additional_trust_roots: Vec<AdditionalTrustRoot>,
+
     /// Maximum lifetime the created certificates are allowed to have.
     /// In case consumers request a longer lifetime than allowed by this setting,
     /// the lifetime will be the minimum of both, so this setting takes precedence.
@@ -135,6 +139,17 @@ impl AutoTlsCa {
     fn default_ca_certificate_lifetime() -> Duration {
         backend::tls::DEFAULT_CA_CERT_LIFETIME
     }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "camelCase")]
+pub struct AdditionalTrustRoot {
+    /// Reference (name and namespace) to a Kubernetes Secret object where additional certificates
+    /// are stored.
+    /// The extensions of the keys denote its contents: A key suffixed with `.pem` contains a stack
+    /// of base64 encoded DER certificates, a key suffixed with `.cer`, `.cert`, or `.crt` contains
+    /// either a binary DER certificate or a stack of base64 encoded DER certificates.
+    pub secret: SecretReference,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -417,6 +432,7 @@ mod test {
                             length: CertificateKeyGeneration::RSA_KEY_LENGTH_3072
                         }
                     },
+                    additional_trust_roots: vec![],
                     max_certificate_lifetime: DEFAULT_MAX_CERT_LIFETIME,
                 })
             }
@@ -436,6 +452,10 @@ mod test {
                   namespace: default
                 autoGenerate: true
                 caCertificateLifetime: 100d
+              additionalTrustRoots:
+                - secret:
+                    name: tls-root-ca
+                    namespace: default
               maxCertificateLifetime: 31d
         "#;
         let deserializer = serde_yaml::Deserializer::from_str(input);
@@ -454,6 +474,12 @@ mod test {
                         ca_certificate_lifetime: Duration::from_days_unchecked(100),
                         key_generation: CertificateKeyGeneration::default()
                     },
+                    additional_trust_roots: vec![AdditionalTrustRoot {
+                        secret: SecretReference {
+                            name: "tls-root-ca".to_string(),
+                            namespace: "default".to_string(),
+                        }
+                    }],
                     max_certificate_lifetime: Duration::from_days_unchecked(31),
                 })
             }